requirements.txt
discord.py
pyarango
pytest
pytest-azurepipelines
pytest-cov
